feat: Update client address to use secure connection
Update the client address in the client.go file to use a secure connection with the eventize.zj.117503445.top server. This ensures that the client communicates securely with the server.

Also, remove the commented out insecure credentials import and replace it with the necessary code to establish a secure connection using TLS.

Lastly, add Dockerfile and note.md files for building and pushing the eventize application to a container registry.

Signed-off-by: 117503445 <t117503445@gmail.com>